requirements.txt
Flask==1.1.2
numpy==1.20.3
tensorflow-cpu==2.5.0
opencv-python-headless==4.5.3.56
Pillow==8.3.2
gunicorn==20.1.0
